+++ doc/changelog.rst
@@ -5,6 +5,16 @@
 Changelog
 =========
 
+2.2.0.rc1
+---------
+
+New features
+~~~~~~~~~~~~
+- Cost functions in iminuit.cost are now additive, creating a new cost function with the
+  union of parameters that returns the sum of the results of the individual cost functions
+- iminuit.cost.NormalConstraint was added as a means to add soft constraints on a
+  parameter, can also be used to set up a covariance matrix between several parameters
+
 2.1.0 (December 18, 2020)
 -------------------------